moved commands that use rpcClient to node command
