+++ cmd/skywire-cli/commands/node/routes.go
@@ -1,4 +1,4 @@
-package rtfind
+package node
 
 import (
 	"errors"
@@ -13,33 +13,25 @@ import (
 
 	"github.com/skycoin/skywire/cmd/skywire-cli/internal"
 
-	"github.com/skycoin/skywire/pkg/cipher"
 	"github.com/skycoin/skywire/pkg/node"
-	"github.com/skycoin/skywire/pkg/route-finder/client"
 	"github.com/skycoin/skywire/pkg/router"
 	"github.com/skycoin/skywire/pkg/routing"
 )
 
-// RtFindCmd contains commands that interact with the route finder
-var RtFindCmd = &cobra.Command{
-	Use:   "rtfind",
-	Short: "Commands that interact with the route-finder",
-}
-
 func init() {
-	RtFindCmd.AddCommand(
+	RootCmd.AddCommand(
 		listRulesCmd,
 		ruleCmd,
 		rmRuleCmd,
 		addRuleCmd,
-		findRoutesCmd)
+	)
 }
 
 var listRulesCmd = &cobra.Command{
 	Use:   "list-rules",
 	Short: "lists the local node's routing rules",
 	Run: func(_ *cobra.Command, _ []string) {
-		rules, err := internal.RPCClient().RoutingRules()
+		rules, err := rpcClient().RoutingRules()
 		internal.Catch(err)
 
 		printRoutingRules(rules...)
@@ -54,7 +46,7 @@ var ruleCmd = &cobra.Command{
 		id, err := strconv.ParseUint(args[0], 10, 32)
 		internal.Catch(err)
 
-		rule, err := internal.RPCClient().RoutingRule(routing.RouteID(id))
+		rule, err := rpcClient().RoutingRule(routing.RouteID(id))
 		internal.Catch(err)
 
 		printRoutingRules(&node.RoutingEntry{Key: rule.RouteID(), Value: rule})
@@ -68,7 +60,7 @@ var rmRuleCmd = &cobra.Command{
 	Run: func(_ *cobra.Command, args []string) {
 		id, err := strconv.ParseUint(args[0], 10, 32)
 		internal.Catch(err)
-		internal.Catch(internal.RPCClient().RemoveRoutingRule(routing.RouteID(id)))
+		internal.Catch(rpcClient().RemoveRoutingRule(routing.RouteID(id)))
 		fmt.Println("OK")
 	},
 }
@@ -117,38 +109,10 @@ var addRuleCmd = &cobra.Command{
 			)
 			rule = routing.ForwardRule(time.Now().Add(expire), nextRouteID, nextTpID)
 		}
-		rIDKey, err := internal.RPCClient().AddRoutingRule(rule)
+		rIDKey, err := rpcClient().AddRoutingRule(rule)
 		internal.Catch(err)
 		fmt.Println("Routing Rule Key:", rIDKey)
 	},
-}
-
-var frAddr string
-var frMinHops, frMaxHops uint16
-
-var findRoutesCmd = &cobra.Command{
-	Use:   "find-routes <public-key-node-1> <public-key-node-2>",
-	Short: "lists available routes between two nodes via route finder service",
-	Args:  cobra.MinimumNArgs(2),
-	Run: func(_ *cobra.Command, args []string) {
-		rfc := client.NewHTTP(frAddr)
-
-		var srcPK, dstPK cipher.PubKey
-		internal.Catch(srcPK.Set(args[0]))
-		internal.Catch(dstPK.Set(args[1]))
-
-		forward, reverse, err := rfc.PairedRoutes(srcPK, dstPK, frMinHops, frMaxHops)
-		internal.Catch(err)
-
-		fmt.Println("forward: ", forward)
-		fmt.Println("reverse: ", reverse)
-	},
-}
-
-func init() {
-	findRoutesCmd.Flags().StringVar(&frAddr, "addr", "https://routefinder.skywire.skycoin.net", "address in which to contact route finder service")
-	findRoutesCmd.Flags().Uint16Var(&frMinHops, "min-hops", 1, "min hops for the returning routeFinderRoutesCmd")
-	findRoutesCmd.Flags().Uint16Var(&frMaxHops, "max-hops", 1000, "max hops for the returning routeFinderRoutesCmd")
 }
 
 func printRoutingRules(rules ...*node.RoutingEntry) {
